updated franklin
