+++ .github/workflows/semantic-release.yaml
@@ -6,13 +6,13 @@ on:
       - main
 jobs:
   semantic-release:
-    if: "!contains(github.event.head_commit.message, 'Initial commit')"
+    if: ${{ ! contains(github.event.head_commit.message,  'Initial commit') }}
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
       - name: Setup Node.js
-        uses: actions/setup-node@v2
+        uses: actions/setup-node@v4
         with:
           node-version: 18
       - name: Install dependencies
@@ -21,5 +21,3 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: npm run semantic-release
-      - name: Show working copy
-        run: ls -la && git status
