+++ .renovaterc.json
@@ -1,3 +1,3 @@
 {
-  "extends": ["github>adobe/helix-shared"]
+  "extends": ["config:recommended"]
 }
