+++ CHANGELOG.md
@@ -1,35 +1,58 @@
-## [1.2.2](https://github.com/elc9aya2ls612j/helix-project-boilerplate/compare/v1.2.1...v1.2.2) (2023-06-29)
+# [1.3.0](https://github.com/adobe/aem-boilerplate/compare/v1.2.2...v1.3.0) (2023-11-30)
 
 
 ### Bug Fixes
 
-* **github:** use double quotes ([d7e285d](https://github.com/elc9aya2ls612j/helix-project-boilerplate/commit/d7e285dbb452617a5166569f91da4cb376548a38))
-
-## [1.2.1](https://github.com/elc9aya2ls612j/helix-project-boilerplate/compare/v1.2.0...v1.2.1) (2023-06-29)
-
-
-### Bug Fixes
-
-* **github:** create and ignore bak files ([ac087b6](https://github.com/elc9aya2ls612j/helix-project-boilerplate/commit/ac087b61d3e44910d980b0d9e7aead0f2ed83873))
-
-# [1.2.0](https://github.com/elc9aya2ls612j/helix-project-boilerplate/compare/v1.1.0...v1.2.0) (2023-06-29)
+* attempt to fix release ([5bb8f73](https://github.com/adobe/aem-boilerplate/commit/5bb8f7302c70e98a60fd9b8a38f63577f90d7414))
+* **build:** fix org in issues URL ([7226d84](https://github.com/adobe/aem-boilerplate/commit/7226d8465e0f5d3c2f4c49bacfbc758c6da5f823))
+* **build:** no double slashes in issues url ([1191064](https://github.com/adobe/aem-boilerplate/commit/119106489e1bfb92c80aa55cc2b634496915cfc1))
+* **github:** increase default permissions for cleanup action ([f8ed051](https://github.com/adobe/aem-boilerplate/commit/f8ed05169e841a21f1a17c664ace76942d7700c8))
+* **github:** restrict running of action to initial commit on main branch ([1a990c2](https://github.com/adobe/aem-boilerplate/commit/1a990c25a4132a50e8c73f80c57845d7d9657399))
+* icons no more svgs ([#267](https://github.com/adobe/aem-boilerplate/issues/267)) ([2d4cfa7](https://github.com/adobe/aem-boilerplate/commit/2d4cfa73a9c35c065e9dd2096a9bab3637ddd9b4))
+* **lib:** update scripts/aem.js to aem.js@1.3.2 ([#266](https://github.com/adobe/aem-boilerplate/issues/266)) ([2d61c6a](https://github.com/adobe/aem-boilerplate/commit/2d61c6a9fb0905159a568abfe73f4d09dfc0846a))
+* rescope icon identifiers to avoid clashing references across icons ([#241](https://github.com/adobe/aem-boilerplate/issues/241)) ([e8dc533](https://github.com/adobe/aem-boilerplate/commit/e8dc53350de838a7dfa094bbbf62c80bdaffdca2)), closes [#235](https://github.com/adobe/aem-boilerplate/issues/235)
+* sampleRUM always for checkpoint to be called even if RUM not selected ([fcca39d](https://github.com/adobe/aem-boilerplate/commit/fcca39dd4f5fd2aef6852580873ab4b2cce1e2af))
+* trigger release ([afbd201](https://github.com/adobe/aem-boilerplate/commit/afbd201189b13a238ad4709923a5ad9b94dc73eb))
 
 
 ### Features
 
-* **github:** populate pull request template ([4ccb759](https://github.com/elc9aya2ls612j/helix-project-boilerplate/commit/4ccb7592a84fdc0d3d234fc4da5bdf94a026ff1d))
+* add capability to patch block config at runtime ([#246](https://github.com/adobe/aem-boilerplate/issues/246)) ([a774acc](https://github.com/adobe/aem-boilerplate/commit/a774acc651359137422ea734d014164509ae5940))
+* enable listeners to all RUM events, not just the sampled ones ([871ede4](https://github.com/adobe/aem-boilerplate/commit/871ede401d2d57c8825f8970f3b28cd9de5f27f8))
+* use aem.js ([#255](https://github.com/adobe/aem-boilerplate/issues/255)) ([f5ab4df](https://github.com/adobe/aem-boilerplate/commit/f5ab4df26a992f70a123f56b08062f7dc0b47dae))
 
-# [1.1.0](https://github.com/elc9aya2ls612j/helix-project-boilerplate/compare/v1.0.5...v1.1.0) (2023-06-29)
-
-
-### Features
-
-* **github:** remove dependencies upon fork, fill in readme ([7e60519](https://github.com/elc9aya2ls612j/helix-project-boilerplate/commit/7e60519c8ea97640bcb064cb3592990989fe10ef))
-
-## [1.0.5](https://github.com/elc9aya2ls612j/helix-project-boilerplate/compare/v1.0.4...v1.0.5) (2023-06-29)
+## [1.2.2](https://github.com/elc9aya2ls612j/aem-boilerplate/compare/v1.2.1...v1.2.2) (2023-06-29)
 
 
 ### Bug Fixes
 
-* trigger release ([2c3244f](https://github.com/elc9aya2ls612j/helix-project-boilerplate/commit/2c3244fc181fd900293bbd7e67ab8e68e5d83d5d))
-* trigger release ([5ac78b0](https://github.com/elc9aya2ls612j/helix-project-boilerplate/commit/5ac78b07955c0b75a37dfb293a5d616b7bdaffba))
+* **github:** use double quotes ([d7e285d](https://github.com/elc9aya2ls612j/aem-boilerplate/commit/d7e285dbb452617a5166569f91da4cb376548a38))
+
+## [1.2.1](https://github.com/elc9aya2ls612j/aem-boilerplate/compare/v1.2.0...v1.2.1) (2023-06-29)
+
+
+### Bug Fixes
+
+* **github:** create and ignore bak files ([ac087b6](https://github.com/elc9aya2ls612j/aem-boilerplate/commit/ac087b61d3e44910d980b0d9e7aead0f2ed83873))
+
+# [1.2.0](https://github.com/elc9aya2ls612j/aem-boilerplate/compare/v1.1.0...v1.2.0) (2023-06-29)
+
+
+### Features
+
+* **github:** populate pull request template ([4ccb759](https://github.com/elc9aya2ls612j/aem-boilerplate/commit/4ccb7592a84fdc0d3d234fc4da5bdf94a026ff1d))
+
+# [1.1.0](https://github.com/elc9aya2ls612j/aem-boilerplate/compare/v1.0.5...v1.1.0) (2023-06-29)
+
+
+### Features
+
+* **github:** remove dependencies upon fork, fill in readme ([7e60519](https://github.com/elc9aya2ls612j/aem-boilerplate/commit/7e60519c8ea97640bcb064cb3592990989fe10ef))
+
+## [1.0.5](https://github.com/elc9aya2ls612j/aem-boilerplate/compare/v1.0.4...v1.0.5) (2023-06-29)
+
+
+### Bug Fixes
+
+* trigger release ([2c3244f](https://github.com/elc9aya2ls612j/aem-boilerplate/commit/2c3244fc181fd900293bbd7e67ab8e68e5d83d5d))
+* trigger release ([5ac78b0](https://github.com/elc9aya2ls612j/aem-boilerplate/commit/5ac78b07955c0b75a37dfb293a5d616b7bdaffba))
