+++ blocks/collection-cards/collection-cards.css
@@ -8,10 +8,6 @@
   display: grid;
   grid-template-columns: repeat(3, minmax(0, 1fr));
   grid-gap: 2rem;
-
-  @media (max-width: 768px) {
-    grid-template-columns: repeat(1, minmax(0, 1fr));
-  }
 }
 
 .collection-cards .collection {
@@ -23,10 +19,6 @@
   border-radius: 0.5rem;
   aspect-ratio: 5 / 6;
   object-fit: cover;
-
-  @media (max-width: 768px) {
-    aspect-ratio: 3 / 2;
-  }
 }
 
 .collection-cards .collection:hover img {
@@ -50,4 +42,14 @@
   pointer-events: none;
   color: gray;
   font-weight: 400;
+}
+
+@media (max-width: 768px) {
+  .collection-cards {
+    grid-template-columns: repeat(1, minmax(0, 1fr));
+  }
+
+  .collection-cards img {
+    aspect-ratio: 3 / 2;
+  }
 }
