+++ blocks/container/container.css
@@ -4,8 +4,10 @@
   border-radius: 0.5rem;
   text-align: center;
   padding: 8rem 10rem;
+}
 
-  @media (max-width: 768px) {
+@media (max-width: 768px) {
+  .container-container .container.support {
     padding: 8rem 2rem;
   }
 }
