+++ blocks/container/container.js
@@ -1,0 +1,2 @@
+export default function decorate() {
+}
