+++ blocks/footer/footer.css
@@ -15,11 +15,6 @@ footer .content {
   padding: 5rem 0;
   display: grid;
   grid-template-columns: repeat(4, minmax(0, 1fr));
-
-  @media (max-width: 768px) {
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-    row-gap: 1rem;
-  }
 }
 
 footer .footer p {
@@ -49,4 +44,11 @@ footer .copyright>div {
   display: flex;
   justify-content: flex-end;
   width: 100%;
+}
+
+@media (max-width: 768px) {
+  footer .content {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+    row-gap: 1rem;
+  }  
 }
