+++ blocks/header/header.js
@@ -1,5 +1,5 @@
 import { getCart, getCartId } from '../../scripts/cart.js';
-import { getMetadata, decorateIcons } from '../../scripts/lib-franklin.js';
+import { getMetadata, decorateIcons } from '../../scripts/aem.js';
 import createTag from '../../utils/tag.js';
 
 // media query match that indicates mobile/tablet width
@@ -10,7 +10,7 @@ function closeOnEscape(e) {
     const nav = document.getElementById('nav');
     const navSections = nav.querySelector('.nav-sections');
     const navSectionExpanded = navSections.querySelector(
-      '[aria-expanded="true"]'
+      '[aria-expanded="true"]',
     );
     if (navSectionExpanded && isDesktop.matches) {
       // eslint-disable-next-line no-use-before-define
@@ -57,20 +57,19 @@ function toggleAllNavSections(sections, expanded = false) {
  * @param {*} forceExpanded Optional param to force nav expand behavior when not null
  */
 function toggleMenu(nav, navSections, forceExpanded = null) {
-  const expanded =
-    forceExpanded !== null
-      ? !forceExpanded
-      : nav.getAttribute('aria-expanded') === 'true';
+  const expanded = forceExpanded !== null
+    ? !forceExpanded
+    : nav.getAttribute('aria-expanded') === 'true';
   const button = nav.querySelector('.nav-hamburger button');
   document.body.style.overflowY = expanded || isDesktop.matches ? '' : 'hidden';
   nav.setAttribute('aria-expanded', expanded ? 'false' : 'true');
   toggleAllNavSections(
     navSections,
-    expanded || isDesktop.matches ? 'false' : 'true'
+    expanded || isDesktop.matches ? 'false' : 'true',
   );
   button.setAttribute(
     'aria-label',
-    expanded ? 'Open navigation' : 'Close navigation'
+    expanded ? 'Open navigation' : 'Close navigation',
   );
   // enable nav dropdown keyboard accessibility
   const navDrops = navSections.querySelectorAll('.nav-drop');
@@ -124,14 +123,14 @@ export default async function decorate(block) {
       if (c === 'banner' && section) {
         block.appendChild(section);
         header.style.height = getComputedStyle(
-          document.documentElement
+          document.documentElement,
         ).getPropertyValue('--extended-nav-height');
       } else if (c === 'brand' && section) {
         const logoLink = section.querySelector('a');
         logoLink.setAttribute('aria-label', 'Home');
       } else {
         header.style.height = getComputedStyle(
-          document.documentElement
+          document.documentElement,
         ).getPropertyValue('--nav-height');
       }
     });
@@ -139,16 +138,14 @@ export default async function decorate(block) {
     const navSections = nav.querySelector('.nav-sections');
     if (navSections) {
       navSections.querySelectorAll(':scope > ul > li').forEach((navSection) => {
-        if (navSection.querySelector('ul'))
-          navSection.classList.add('nav-drop');
+        if (navSection.querySelector('ul')) { navSection.classList.add('nav-drop'); }
         navSection.addEventListener('click', () => {
           if (isDesktop.matches) {
-            const expanded =
-              navSection.getAttribute('aria-expanded') === 'true';
+            const expanded = navSection.getAttribute('aria-expanded') === 'true';
             toggleAllNavSections(navSections);
             navSection.setAttribute(
               'aria-expanded',
-              expanded ? 'false' : 'true'
+              expanded ? 'false' : 'true',
             );
           }
         });
@@ -166,9 +163,7 @@ export default async function decorate(block) {
     nav.setAttribute('aria-expanded', 'false');
     // prevent mobile nav behavior on window resize
     toggleMenu(nav, navSections, isDesktop.matches);
-    isDesktop.addEventListener('change', () =>
-      toggleMenu(nav, navSections, isDesktop.matches)
-    );
+    isDesktop.addEventListener('change', () => toggleMenu(nav, navSections, isDesktop.matches));
 
     await decorateIcons(nav);
