+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@adobe/helix-project-boilerplate",
+  "name": "@adobe/aem-boilerplate",
   "private": true,
-  "version": "1.2.2",
+  "version": "1.3.0",
   "description": "Starter project for Adobe Helix",
   "scripts": {
     "lint:js": "eslint .",
@@ -11,30 +11,26 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/adobe/helix-project-boilerplate.git"
+    "url": "git+https://github.com/adobe/aem-boilerplate.git"
   },
   "author": "Adobe",
   "license": "Apache License 2.0",
   "bugs": {
-    "url": "https://github.com/adobe/helix-project-boilerplate/issues"
+    "url": "https://github.com/adobe/aem-boilerplate/issues"
   },
-  "homepage": "https://github.com/adobe/helix-project-boilerplate#readme",
+  "homepage": "https://github.com/adobe/aem-boilerplate#readme",
   "devDependencies": {
     "@semantic-release/changelog": "6.0.3",
     "@semantic-release/exec": "6.0.3",
     "@semantic-release/git": "10.0.1",
-    "semantic-release": "21.0.5",
-    "@babel/core": "7.21.0",
-    "@babel/eslint-parser": "7.19.1",
-    "chai": "4.3.7",
-    "eslint": "8.35.0",
+    "semantic-release": "23.0.0",
+    "@babel/core": "7.23.7",
+    "@babel/eslint-parser": "7.23.3",
+    "chai": "5.0.0",
+    "eslint": "8.56.0",
     "eslint-config-airbnb-base": "15.0.0",
-    "eslint-plugin-import": "2.27.5",
-    "@esm-bundle/chai": "4.3.4-fix.0",
-    "@web/test-runner": "0.15.1",
-    "@web/test-runner-commands": "0.6.5",
-    "sinon": "15.0.1",
-    "stylelint": "15.2.0",
-    "stylelint-config-standard": "30.0.1"
+    "eslint-plugin-import": "2.29.1",
+    "stylelint": "16.1.0",
+    "stylelint-config-standard": "36.0.0"
   }
 }
