+++ scripts/config.js
@@ -1,8 +1,6 @@
-const STRIPE_PK =
-  'pk_test_51Nrs0ML1DV9f5cSo8SGP5fqOt9ypRLqBSKBv4rfbRPOGezL5t5sUMV7mHGlRdd9455BXOwoBngtEWo35EEgc7UHD00Ajy9FWgK';
+const STRIPE_PK = 'pk_test_51Nrs0ML1DV9f5cSo8SGP5fqOt9ypRLqBSKBv4rfbRPOGezL5t5sUMV7mHGlRdd9455BXOwoBngtEWo35EEgc7UHD00Ajy9FWgK';
 const GRAPHQL_ENDPOINT = 'http://localhost/graphql';
-const CLIENT_SECRET =
-  'pi_3NsZkUL1DV9f5cSo0t4biUBY_secret_hbaemymPGMHsuf8oxqMYthy2p';
+const CLIENT_SECRET = 'pi_3NsZkUL1DV9f5cSo0t4biUBY_secret_hbaemymPGMHsuf8oxqMYthy2p';
 
 const defaultConfigValue = {
   stripe_pk: STRIPE_PK,
@@ -14,12 +12,13 @@ const getEnvironment = () => {
   const { href } = window.location;
   if (href.includes('localhost')) {
     return 'dev';
-  } else if (href.includes('.hlx.page')) {
+  } if (href.includes('.hlx.page')) {
     return 'preview';
   }
   return 'prod';
 };
 
+// eslint-disable-next-line import/prefer-default-export
 export const getConfigValue = (key) => {
   const env = getEnvironment();
   if (env === 'dev') {
