+++ styles/styles.css
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-:root {
+ :root {
   /* colors */
   --link-color: #3C9BAA;
   --link-hover-color: #3C9BAA;
@@ -34,32 +34,20 @@
   --fixed-font-family: 'Roboto Mono', menlo, consolas, 'Liberation Mono', monospace;
 
   /* body sizes */
-  --body-font-size-xxl: 1.5rem;
-  /* 24px */
-  --body-font-size-xl: 1.25rem;
-  /* 20px */
-  --body-font-size-l: 1.125rem;
-  /* 18px */
-  --body-font-size-m: 1rem;
-  /* 16px */
-  --body-font-size-s: 0.875rem;
-  /* 14px */
-  --body-font-size-xs: 0.75rem;
-  /* 12px */
+  --body-font-size-xxl: 24px;   
+  --body-font-size-xl: 20px;   
+  --body-font-size-l: 18px;  
+  --body-font-size-m: 16px;
+  --body-font-size-s: 14px;
+  --body-font-size-xs: 12px;
 
   /* heading sizes */
-  --heading-font-size-xxl: 3rem;
-  /* 48px */
-  --heading-font-size-xl: 2.5rem;
-  /* 40px */
-  --heading-font-size-l: 2rem;
-  /* 32px */
-  --heading-font-size-m: 1.5rem;
-  /* 24px */
-  --heading-font-size-s: 1.25rem;
-  /* 20px */
-  --heading-font-size-xs: 1.125rem;
-  /* 18px */
+  --heading-font-size-xxl: 48px;
+  --heading-font-size-xl: 40px;
+  --heading-font-size-l: 32px;
+  --heading-font-size-m: 24px;
+  --heading-font-size-s: 20px;
+  --heading-font-size-xs: 18px;
 
   /* nav height */
   --nav-height: 64px;
